Adicionadas exceçoes

+++ src/main/java/com/GabrielOliveira/DScommerce/Services/ProductService.java
@@ -1,13 +1,18 @@
 package com.GabrielOliveira.DScommerce.Services;
 
 import com.GabrielOliveira.DScommerce.Repository.ProductRepository;
+import com.GabrielOliveira.DScommerce.Services.Exception.DataBaseException;
+import com.GabrielOliveira.DScommerce.Services.Exception.ResourceNotFoundException;
 import com.GabrielOliveira.DScommerce.dto.ProductDTO;
 import com.GabrielOliveira.DScommerce.entities.Product;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
 
 @Service
 public class ProductService {
@@ -16,46 +21,71 @@ public class ProductService {
     private ProductRepository repository;
 
     @Transactional(readOnly = true)
-    public ProductDTO findById(Long id){
+    public ProductDTO findById(Long id) {
 
-        return new ProductDTO(repository.findById(id).orElseThrow());
+        return new ProductDTO(repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado")));
     }
 
 
     @Transactional(readOnly = true)
-    public Page<ProductDTO> findAll(Pageable page){
+    public Page<ProductDTO> findAll(Pageable page) {
         return repository.findAll(page).map(x -> new ProductDTO(x));
     }
 
-    public ProductDTO insert(ProductDTO dto){
+    public ProductDTO insert(ProductDTO dto) {
         return new ProductDTO(repository.save(new Product(dto)));
     }
 
     @Transactional
-    public ProductDTO update(Long id, ProductDTO dto){
+    public ProductDTO update(Long id, ProductDTO dto) {
 
-       Product entity = repository.getReferenceById(id);
 
-       dtoToEntity(dto, entity);
+        try {
 
-       repository.save(entity);
+            Product entity = repository.getReferenceById(id);
 
-       return new ProductDTO(entity);
+            dtoToEntity(dto, entity);
+
+            repository.save(entity);
+
+            return new ProductDTO(entity);
+
+        } catch (Exception e) {
+            throw new ResourceNotFoundException("Recurso nao encontrado");
+        }
+
+    }
+
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso nao entrado");
+        }
+
+        try {
+
+            repository.deleteById(id);
+
+        } catch (Exception e) {
+
+            throw new DataBaseException("Falha de integridade referencial");
+
+        }
 
     }
 
     @Transactional
-    public void delete(Long id){
-        repository.deleteById(id);
+    public List<ProductDTO> findByPriceLessThan(Double price) {
+        return repository.findByPriceLessThan(price).stream().map(x -> new ProductDTO(x)).toList();
     }
 
-    private void dtoToEntity(ProductDTO dto, Product entity){
+    private void dtoToEntity(ProductDTO dto, Product entity) {
         entity.setName(dto.getName());
         entity.setDescription(dto.getDescription());
         entity.setPrice(dto.getPrice());
         entity.setImgUrl(dto.getImgUrl());
     }
-
 
 
 }
